[MISC] move type_traits/concept to utility
